포스트맨으로 , DELETE, EDIT 요청 보내면, 미인증 사용자도 아무 상품이나 DELETE 또는 EDIT 가능함. 우선 표시만 예외 표시만 해뒀음 (실제로는 안터짐, 공부하고 수정할 예정) + 주석을 달아놈. 검증로직 이후 [성공] 아래부터 성공이후 로직이다.

+++ src/main/java/demo/demo/controller/ProductController.java
@@ -31,6 +31,7 @@ public class ProductController {
 
     @GetMapping("")
     public String getProducts(Model model) {
+        //[성공]
         List<Product> products = productService.findAllProduct();
         model.addAttribute("products", products);
         return "products/productsList";
@@ -41,9 +42,13 @@ public class ProductController {
     @GetMapping("/new")
     public String getCreateView(@SessionAttribute(name = SessionConst.LOGIN_USER,  required = false) User loginUser,
                                 Model model) {
+        // 미인증 사용자는 상품생성 불가
         if (loginUser == null) {
             return "redirect:/login";
         }
+
+        // [성공]
+        // 상품생성 뷰 이동
         model.addAttribute("product", new ProductCreateForm());
         return "products/product-new";
     }
@@ -53,15 +58,19 @@ public class ProductController {
                          BindingResult bindingResult,
                          @SessionAttribute(SessionConst.LOGIN_USER) User loginUser,
                          RedirectAttributes redirectAttributes) {
+        // 미인증 사용자
         if (loginUser == null) {
             return "redirect:/";
         }
 
+        // 검증실패
         if (bindingResult.hasErrors()) {
             log.info("{}", bindingResult);
             return "products/product-new";
         }
 
+        // [성공]
+        // 상품생성 성공 후, 상품상세 뷰로 이동
         Product product = productService.create(form, loginUser.getId());
         redirectAttributes.addAttribute("productId", product.getId());
         return "redirect:/products/{productId}";
@@ -73,28 +82,36 @@ public class ProductController {
                                   @SessionAttribute(name = SessionConst.LOGIN_USER) User loginUser,
                               RedirectAttributes redirectAttributes,
                               Model model) {
-
         Product product = productService.findById(productId).orElseThrow(NoSuchElementException::new);
-
-
         // 상품 주인이 아닌 경우
-        if (!productService.ownerCheck(loginUser.getId(), product.getOwnerId())) {
+        if (!productService.ownerCheck(loginUser.getId(), product)) {
             redirectAttributes.addAttribute("productId", product.getId());
             return "redirect:/products/{productId}";
         }
 
+        // [성공]
+        // 상품 수정 폼 생성 후, 모델에 추가
         ProductUpdateForm productUpdateForm = productService.updateProductCreate(product);
         model.addAttribute("product", productUpdateForm);
         model.addAttribute("createDate", product.getCreateDate());
+        // 상품 이미지들 모델에 추가
         List<String> imageFileNames = productService.getImageFileNames(product);
         model.addAttribute("imageFileNames", imageFileNames);
+        // 상품수정 뷰 이동
         return "products/product-edit";
     }
     @PostMapping("/edit/{id}")
     public String updateProduct(@Validated @ModelAttribute("product") ProductUpdateForm form,
                          BindingResult bindingResult,
+                         @SessionAttribute(name = SessionConst.LOGIN_USER) User loginUser,
                          RedirectAttributes redirectAttributes,
                                 Model model) {
+        // 미인증 사용자 요청
+        if (loginUser == null) {
+            throw new IllegalStateException();
+        }
+
+        // 검증 실패 -> 상품이미지와, 상품수정일 리다이렉트
         if (bindingResult.hasErrors()) {
             Product product = productService.findById(form.getId()).orElseThrow(NoSuchElementException::new);
             model.addAttribute("createDate", product.getCreateDate());
@@ -102,6 +119,8 @@ public class ProductController {
             return "products/product-edit";
         }
 
+        // [성공]
+        // 상품수정 성공
         redirectAttributes.addAttribute("productId", productService.update(form));
         return "redirect:/products/{productId}";
     }
@@ -112,25 +131,47 @@ public class ProductController {
                           Model model,
                           @SessionAttribute(name = SessionConst.LOGIN_USER, required = false)  User loginUser) {
         Product product = productService.findById(productId).orElse(null);
+        // 상품이 없는 경우
         if (product == null) {
+            //throw new RuntimeException();
             return "errors/product-null";
         }
 
+        // [성공]
+        // 상품 모델에 넣기
         model.addAttribute("product", product);
+        // 상품 이미지들 모델에 넣기
         List<String> imageFileNames = productService.getImageFileNames(product);
         model.addAttribute("imageFileNames", imageFileNames);
 
+        // 로그인한 유저인 경우, [로그인유저 아이디][상품주인 아이디]를 비교해 상품 수정버튼 시각적 여부 확인
         if (loginUser != null) {
             model.addAttribute("loginUserId", loginUser.getId());
         }
+        // 상품상세 뷰 이동
         return "products/product-view";
     }
 
     @DeleteMapping("/{productId}")
-    public ResponseEntity<String> deleteProduct(@PathVariable Long productId) {
+    public ResponseEntity<String> deleteProduct(@PathVariable Long productId,
+                                                @SessionAttribute(name = SessionConst.LOGIN_USER) User loginUser) {
+        Product product = productService.findById(productId).orElse(null);
+        // 상품이 없는 경우
+        if (product == null) {
+            throw new IllegalStateException();
+        }
+
+        // 상품 주인이 아닌 경우
+        if (!(productService.ownerCheck(loginUser.getId(), product))) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("상품 삭제에 실패했습니다.");
+        }
+
+        // 디비에서 상품 삭제 실패한 경우
         if (!productService.delete(productId)) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("상품 삭제에 실패했습니다.");
         }
+
+        // [성공] 상품삭제 성공 응답
         return ResponseEntity.ok("상품이 삭제되었습니다.");
     }
